Add: Store the last position when follow a link and go back
When following a link, the position is stored. A button is
displayed and allow the user to go back to the position
before following the link. This allow to continue reading
after jumping to a footnote by example.

When the user click a link, we check if the link destination is the
same position we stored just before. By example, after jumping
to a footnote, if the user click the link which allow to return
to the mention. If positions are similar, we do not follow the link
but we go back using the stored position before going to the footnote.